Tempo rambo - more bbt scale fiddling.

+++ gtk2_ardour/tempo_lines.cc
@@ -77,7 +77,7 @@ TempoLines::draw_ticks (std::vector<ARDOUR::TempoMap::BBTPoint>& grid,
 		framepos_t f = 0;
 
 		if (grid.begin()->c != 0.0) {
-			const double pulses_per_div = l * (grid.begin()->tempo.note_type() / grid.begin()->meter->note_divisor()) / divisions;
+			const double pulses_per_div = l * (grid.begin()->tempo.note_type() / grid.begin()->meter->note_divisor()) / (double) divisions;
 			const double time_at_pulse = log (((grid.begin()->c * (pulses_per_div / grid.begin()->tempo.note_type())) /
 							   grid.begin()->tempo.pulses_per_minute()) + 1) / grid.begin()->c;
 			f = grid.begin()->frame + (framecnt_t) floor ((time_at_pulse * 60.0 * frame_rate) + 0.5);
